First cut at intro page for hurricane data hacking
For the moment, I focused on removing mentions of oceanographic data and updating the language to reflect that this will be a modularized version of the hackweek approach.

We also need to think about what are the appropriate first step technical exercises.  For example, should we really have them doing github, jupyterhub, and python at this stage?